fix(ci): python full coverage — sync Grok readiness contracts and split oversized files
