+++ tests/unit/api/test_health_parts/test_health_part_001.py
@@ -280,6 +280,7 @@ async def test_readyz_response_shape_matches_contract(
         "claude_code",
         "gemini",
         "opencode",
+        "grok",
         "docker",
     }
     assert body["agent_readiness"]["security"]["status"] == "warning"
